Daqi alert implementation + OS names error 500 fix + Email alert defect fix
